add authorization for task_statuses

+++ app/Http/Controllers/TaskStatusController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\TaskStatus;
+use Illuminate\Support\Facades\Gate;
 
 class TaskStatusController extends Controller
 {
@@ -15,12 +16,18 @@ class TaskStatusController extends Controller
 
     public function create()
     {
+        if (! Gate::allows('create-task-status')) {
+            abort(403);
+        }
         $taskStatus = new TaskStatus();
         return view('task_statuses.create', compact('taskStatus'));
     }
 
     public function store(Request $request)
     {
+        if (! Gate::allows('create-task-status')) {
+            abort(403);
+        }
         $data = $this->validate($request, [
             'name' => 'required|unique:task_statuses|min:2'
         ]);
@@ -35,6 +42,9 @@ class TaskStatusController extends Controller
 
     public function edit(TaskStatus $taskStatus)
     {
+        if (! Gate::allows('update-task-status')) {
+            abort(403);
+        }
         $id = $taskStatus->id;
         $taskStatus = TaskStatus::find($id);
         return view('task_statuses.edit', compact('taskStatus'));
@@ -42,6 +52,9 @@ class TaskStatusController extends Controller
 
     public function update(Request $request, TaskStatus $taskStatus)
     {
+        if (! Gate::allows('update-task-status')) {
+            abort(403);
+        }
         $id = $taskStatus->id;
         $taskStatus = TaskStatus::findOrFail($id);
         $data = $this->validate($request, [
@@ -56,16 +69,17 @@ class TaskStatusController extends Controller
 
     public function destroy(TaskStatus $taskStatus)
     {
+        if (! Gate::allows('destroy-task-status')) {
+            abort(403);
+        }
         if ($taskStatus) {
             $taskStatus->delete();
         }
         return redirect()->route('task_statuses.index');
     }
 
-    public function show(TaskStatus $taskStatus)
+    public function show()
     {
-        $id = $taskStatus->id;
-        $taskStatus = TaskStatus::findOrFail($id);
         return redirect()->route('task_statuses.index');
     }
 }
